requirements.txt
fastapi
pyjwt
requests
cryptography
uvicorn
